move set prefix
